Refactor: Split event handler into smaller files
Refactor the event handler module into smaller, more manageable files to improve code organization and readability. Remove any unused imports or files after the refactoring is complete.
